requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
coverage==4.1
decorator==4.0.10
defusedxml==0.4.1
Django==1.10
django-allauth==0.26.1
django-debug-toolbar==1.4
django-mptt==0.8.5
django-nose==1.4.4
djangorestframework==3.4.1
ecdsa==0.13
Fabric==1.11.1
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
kombu==3.0.35
nose==1.3.7
oauthlib==1.1.2
paramiko==1.17.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.4
psycopg2==2.6.2
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
python-openid==2.2.5
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.19
traitlets==4.2.2
uWSGI==2.0.13.1
wcwidth==0.1.7
wheel==0.24.0
